backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
aiohttp
google-generativeai
python-dotenv
pypdf